# webservice/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API for managing D\u0026D 5e player characters",
        "title": "Player Character API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8765",
    "basePath": "/",
    "paths": {
        "/api/characters": {
            "get": {
                "description": "Retrieve a paginated list of characters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "List characters",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PaginationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new D\u0026D 5e character with validation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Create a new character",
                "parameters": [
                    {
                        "description": "Character data",
                        "name": "character",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/characters/{id}": {
            "get": {
                "description": "Retrieve a specific character by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Get a character by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing character by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Update a character",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated character data",
                        "name": "character",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ValidationErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a character by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Delete a character",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AbilityScore": {
            "type": "object",
            "properties": {
                "base": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 1
                }
            }
        },
        "models.AbilityScores": {
            "type": "object",
            "required": [
                "charisma",
                "constitution",
                "dexterity",
                "intelligence",
                "strength",
                "wisdom"
            ],
            "properties": {
                "charisma": {
                    "$ref": "#/definitions/models.AbilityScore"
                },
                "constitution": {
                    "$ref": "#/definitions/models.AbilityScore"
                },
                "dexterity": {
                    "$ref": "#/definitions/models.AbilityScore"
                },
                "intelligence": {
                    "$ref": "#/definitions/models.AbilityScore"
                },
                "strength": {
                    "$ref": "#/definitions/models.AbilityScore"
                },
                "wisdom": {
                    "$ref": "#/definitions/models.AbilityScore"
                }
            }
        },
        "models.Character": {
            "type": "object",
            "required": [
                "abilityScores",
                "characterName",
                "class",
                "level",
                "race"
            ],
            "properties": {
                "abilityScores": {
                    "$ref": "#/definitions/models.AbilityScores"
                },
                "alignment": {
                    "type": "string",
                    "enum": [
                        "Lawful",
                        "Good",
                        "Neutral",
                        "Good",
                        "Chaotic",
                        "Good",
                        "Lawful",
                        "Neutral",
                        "True",
                        "Neutral",
                        "Chaotic",
                        "Neutral",
                        "Lawful",
                        "Evil",
                        "Neutral",
                        "Evil",
                        "Chaotic",
                        "Evil"
                    ]
                },
                "background": {
                    "type": "string"
                },
                "characterName": {
                    "type": "string"
                },
                "class": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "experiencePoints": {
                    "type": "integer",
                    "minimum": 0
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 1
                },
                "multiclass": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MulticlassEntry"
                    }
                },
                "playerName": {
                    "type": "string"
                },
                "race": {
                    "type": "string"
                },
                "subclass": {
                    "type": "string"
                },
                "subrace": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.MulticlassEntry": {
            "type": "object",
            "required": [
                "class",
                "level"
            ],
            "properties": {
                "class": {
                    "type": "string"
                },
                "level": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 1
                },
                "subclass": {
                    "type": "string"
                }
            }
        },
        "models.Pagination": {
            "type": "object",
            "properties": {
                "hasNext": {
                    "type": "boolean"
                },
                "hasPrev": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "models.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Character"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/models.Pagination"
                }
            }
        },
        "models.ValidationError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ValidationError"
                    }
                }
            }
        }
    },
    "externalDocs": {
        "description": "OpenAPI",
        "url": "https://swagger.io/resources/open-api/"
    }
}